backend/requirements.txt
Django==4.2
psycopg2==2.9.10
djangorestframework==3.14.0
django-cors-headers==3.14.0
python-dotenv==1.1.0